TaylorGreenVortex initial condition now in flow_solver.cpp, FlowSolver.run_test() can now initialize the flow
